Code cleanup/minor refactoring for npc shuffle item and location pool lists
